requirements.txt
astroquery>=0.4.7
lightkurve>=2.4.0
ellc>=1.8.0
tensorflow>=2.12.0
numpy>=1.23.5
scipy>=1.10.1
matplotlib>=3.7.1
pyvista>=0.38.5
plotly>=5.15.0
dash>=2.11.1
prefect>=2.11.5
scikit-learn>=1.3.0
astropy>=5.3.1
tqdm>=4.65.0
emcee>=3.1.4
corner>=2.2.2
pyyaml>=6.0
argparse>=1.4